+++ src/main/java/vendingmachine/service/OrderService.java
@@ -9,4 +9,5 @@ public class OrderService {
 		cost = OrderException.checkCanTakeThisOrder(orderedMenu, vendingMachine);
 		return cost;
 	}
+
 }
